+++ examples/mlp-vs-knn/script.js
@@ -12,7 +12,7 @@ import {
   modelParameters,
   confidencePlot,
   trainingProgress,
-  textField,
+  textInput,
   toggle,
   trainingPlot,
   webcam,
@@ -27,7 +27,7 @@ import {
 const input = webcam();
 const featureExtractor = mobileNet();
 
-const label = textField();
+const label = textInput();
 label.title = 'Instance label';
 const capture = button({ text: 'Hold to record instances' });
 capture.title = 'Capture instances to the training set';